#1067 Removed OrderItemShipGroup.getEstimatedShipDate calculation, replaced by EECA on updating the related OrderItemShipGrpInvRes
Since estimatedShipDate is now a field in the DB for OrderItemShipGroup, we cannot keep the current code (it provoke an infinite recursion when trying to convert the entity toMap).
Instead we use an EECA which on updating/creating OrderItemShipGrpInvRes will keep the estimatedShipDate up to date.

Note: this need unit tests at some points.

git-svn-id: svn://svn.opentaps.org/opentaps_all/branches/ofbiz-0904-1.4.0-preview1@13167 d3523486-f5fe-0310-a41a-a15a4e76f3c7