#1067 Fix bsh scriplets
For some reason the bsh scriptlets does not work properly and gives class not found exceptions. This seems to happen after the context is loaded in the BSH interpreter as the same expression just before that would evaluate properly.

This is the simplest workaround, changing from bsh: to groovy: interpreter.

git-svn-id: svn://svn.opentaps.org/opentaps_all/branches/ofbiz-0904-1.4.0-preview1@13251 d3523486-f5fe-0310-a41a-a15a4e76f3c7