#1067 Add displayPartyLink macros to easily build a link from a partyId.
Uses org.opentaps.common.party.PartyHelper createViewPageURL and getCrmsfaPartyName, make sure the URL is not escaped but the name is.
Include fix in displayLink as the externalKeyParam ampersand is already HTML escaped and should not be encoded again.

Add org.opentaps.common.party.PartyHelper createViewPageURL that does not need the externalKeyParam, since we use the displayLink macro it will be added by it anyway.

git-svn-id: svn://svn.opentaps.org/opentaps_all/branches/ofbiz-0904-1.4.0-preview1@13425 d3523486-f5fe-0310-a41a-a15a4e76f3c7