#1911 Fix issues in the freemarker template caching.
This seems to be the root cause for the email problems. Since we expand both the body and subject using "MergeForm" as the template location, the first email body would get cached and all following subjects / bodies would have this text as their merged content.

Instead also use the template String as part of the cache key so that different template strings have separate cached values.

git-svn-id: svn://svn.opentaps.org/opentaps_all/versions/1.4/trunk@13870 d3523486-f5fe-0310-a41a-a15a4e76f3c7