#1369 Fix FO error message screen
When the FO screen renderer gets an exception outside of the FO rendering itself, the ofbiz framework inserts an HTML description of the error instead of the XML FO page.
Changed ScreenFopViewHandler not to try to render this HTML as an FO document.
Fixed the FoError screen to be able to display it properly.

Example: /crmsfa/control/shipGroups.pdf?orderId=notexists would trigger this.

git-svn-id: svn://svn.opentaps.org/opentaps_all/versions/1.4/trunk@14773 d3523486-f5fe-0310-a41a-a15a4e76f3c7