Add seed data to make it possible to define a custom footer
Changed the value / defaultValue field type to text, as the previous long-varchar (varchar 255) would be too small if we want to hold HTML fragments (like the footer text).
Using StringUtil.wrapString since the custom footer HTML will most probably include HTML.
Keep the default opentaps footer in the FTL instead of as a default value of the configuration for now because it is currently complicated to get the included FTL code to interpret.