Add was to exclude a GWT module
This is useful when creating a new module that extends an existing module, since in that case GWT will already call the Entry method on the two modules when the extended module is loaded. In that case the base module must be excluded or widgets are duplicated.

Also can be useful when wanting to replace a gwt module in screens imported from another application.