#1838 Fix conflict causing random GWT javascript errors
The problem is that each module is compiled independently and the compiler will remove dead-code (part of the code that is never called in the module).

This is also done in the static JS methods used for widget initialization which is present in each module, but there is a possible conflict as to which module the static method gets called from, so if a module using GWT-ext but no EXT widgets (here webapp) is loaded in the page, other modules will sometime have their widgets call the wrong methods.

Here a workaround is to add a widget in the module causing the code to be present.