Fix bug in the create forecast
Tthe parent period id was set as the period id of the first sub-period instead of it parent period id,
the result is that forecasts were created for each month of a quarter and would appear in the find forecasts which list only quarter forecasts.

After fix a quarter forecast is properly created, which each month as a sub forecast of it.