Change the SECA logic to stop on error
The default behavior would have the all seca being called after an error and the transaction marked for rollback, which causes every seca to fail.
The original service would then have a very big and confusing error message.

This stops the seca as soon as an error occurred, menain the error reported is only the actual cause of the error.

Note: this did not affect any of the opentaps unit test, they are still all passing.